move in-memory event queue from separate namespace (bc break)
