+++ tests/EventBus/CommandBusAwareEventDispatcherTest.phpt
@@ -12,7 +12,7 @@ require_once __DIR__ . '/../bootstrap.php';
 use Damejidlo\EventBus\CommandBusAwareEventDispatcher;
 use Damejidlo\EventBus\IDomainEvent;
 use Damejidlo\EventBus\IEventDispatcher;
-use Damejidlo\EventBus\Implementation\InMemoryEventQueue;
+use Damejidlo\EventBus\InMemoryEventQueue;
 use Damejidlo\MessageBus\Middleware\IsCurrentlyHandlingAwareMiddleware;
 use DamejidloTests\DjTestCase;
 use Mockery;
