+++ tests/Integration/SynchronousEventsDispatchingTest.phpt
@@ -14,7 +14,7 @@ use Damejidlo\CommandBus\ICommandBus;
 use Damejidlo\CommandBus\NewEntityId;
 use Damejidlo\EventBus\CommandBusAwareEventDispatcher;
 use Damejidlo\EventBus\EventBus;
-use Damejidlo\EventBus\Implementation\InMemoryEventQueue;
+use Damejidlo\EventBus\InMemoryEventQueue;
 use Damejidlo\EventBus\SynchronousEventDispatcher;
 use Damejidlo\MessageBus\Handling\HandlerInvokingMiddleware;
 use Damejidlo\MessageBus\Handling\HandlerTypesResolvingMiddleware;
